Resolved RunPage VM issue; changed namespaces to be standard; Registered needed Services.
