+++ RIPDApp/Pages/Run/RunPage.xaml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:vm="clr-namespace:RIPDApp.ViewModels.Run"
-             x:Class="RIPDApp.Pages.Run.RunPage"
+             xmlns:vm="clr-namespace:RIPDApp.ViewModels"
+             x:Class="RIPDApp.Pages.RunPage"
              x:DataType="vm:RunVM"
              Title="RunPage">
