+++ RIPDApp/Services/Run/IRunGpsLocationService.cs
@@ -1,0 +1,5 @@
+﻿namespace RIPDApp.Services;
+
+public interface IRunGpsLocationService
+{
+}
